Addressed the changes requested in PR #9
