+++ src/app/course-list/section/period.ts
@@ -1,0 +1,7 @@
+export class Period {
+  type: string;
+  day: number;
+  // start and end are in MINUTES SINCE START OF THE WEEK // no?
+  start: number;
+  end: number;
+}
