+++ src/app/course-view/component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute, Params } from '@angular/router';
-import { Course } from '../course-list/course/component';
+import { Course } from '../course-list/course/course';
 import { YacsService } from '../services/yacs.service';
 
 
@@ -18,9 +18,12 @@ export class CourseViewComponent implements OnInit {
     private activatedRoute: ActivatedRoute) { }
 
   getCourses (params: Params) {
-    let newParams : Object = {};
-    // add show_sections to params
-    Object.assign(newParams, params, { show_sections: true, show_periods: true }); // cannot directly modify params for some reason
+    let newParams : Object = {
+      show_sections: true,
+      show_periods: true
+    };
+    // add show_sections and show_periods to params
+    Object.assign(newParams, params); // cannot directly modify params
     this.yacsService
         .get('courses', newParams)
         .then((data) => {
