+++ src/app/school-list/department/department.ts
@@ -1,0 +1,5 @@
+export class Department {
+  id: number;
+  code: string;
+  name: string;
+}
