+++ src/app/school-list/school.ts
@@ -1,0 +1,7 @@
+import { Department } from './department/department';
+
+export class School {
+  id: number;
+  name: string;
+  departments: Department[];
+}
